agentic_ai/requirements.txt
# requirements.txt
python-dotenv
langchain
groq
pandas
numpy
click
requests
PyPDF2
openai
